Adding unit testing design pattern
There is a common issue which is that web3 can take some time to perform transactions on the blockchain and then return the appropriate results in time for testing.
Obviously with other languages like Python it is very easy to just make the linear script "sleep". This is not the case with Javascript.
I have decided that the most robust way to create custom unit tests is to create a class structure. I have also created a design pattern which is able to make the application pause (wait/sleep) until the appropriate variables are ready.